Testes  e correções

+++ src/main/java/br/com/dogsteps/repositories/PetRepository.java
@@ -1,12 +1,17 @@
 package br.com.dogsteps.repositories;
 
 import br.com.dogsteps.dao.Dao;
+import br.com.dogsteps.excecoes.StringVaziaException;
+import br.com.dogsteps.excecoes.ValorNegativoException;
 import br.com.dogsteps.interfaces.IDao;
 import br.com.dogsteps.interfaces.IRepository;
 import br.com.dogsteps.models.Pet;
+import org.jetbrains.annotations.NotNull;
+
 import javax.ws.rs.core.Response;
 import java.io.IOException;
 import java.util.List;
+import java.util.Random;
 
 public class PetRepository implements IRepository<Pet, String> {
 
@@ -15,9 +20,9 @@ public class PetRepository implements IRepository<Pet, String> {
 
     @Override
     public Dao inicializarDao() {
-        try{
+        try {
             return new Dao(DIRETORIO);
-        }catch (IOException e) {
+        } catch (IOException e) {
             e.printStackTrace();
             return null;
         }
@@ -30,30 +35,56 @@ public class PetRepository implements IRepository<Pet, String> {
 
     @Override
     public Pet find(String id) {
-        return (Pet) PET_DAO.get(id);
+        return PET_DAO.get(id);
     }
 
     @Override
-    public Response add(Pet pet) {
-        if(PET_DAO.add(pet))
-            return Response.status(Response.Status.OK).build();
-        else
-            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+    public Response add(@NotNull Pet pet) {
+        try {
+            validarRequisicao(pet);
+            if (PET_DAO.add(pet))
+                return Response.status(Response.Status.OK).build();
+            else
+                return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+        } catch (ValorNegativoException | StringVaziaException e) {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        }
     }
 
     @Override
     public Response update(Pet pet) {
-        if(PET_DAO.update(pet))
-            return Response.status(Response.Status.OK).build();
-        else
-            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+        try {
+            validarRequisicao(pet);
+            if (PET_DAO.update(pet))
+                return Response.status(Response.Status.OK).build();
+            else
+                return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+        } catch (ValorNegativoException | StringVaziaException e) {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        }
     }
 
     @Override
     public Response remove(String id) {
-        if(PET_DAO.remove(id))
-            return Response.status(Response.Status.OK).build();
-        else
-            return Response.status((Response.Status.INTERNAL_SERVER_ERROR)).build();
+        if (id != null) {
+            if (!(id.isEmpty())) {
+                if (PET_DAO.remove(id))
+                    return Response.status(Response.Status.OK).build();
+                else
+                    return Response.status(Response.Status.NOT_FOUND).build();
+            }
+        }
+        return Response.status(Response.Status.BAD_REQUEST).build();
+
+    }
+
+    private void validarRequisicao(Pet pet) throws ValorNegativoException, StringVaziaException {
+
+        if (pet.getIdade() <= 0)
+            throw new ValorNegativoException();
+
+        if (pet.getName().isEmpty() || pet.getFoto().isEmpty() || pet.getRaca().isEmpty())
+            throw new StringVaziaException();
+
     }
 }
